requirements.txt
streamlit>=1.24.0
openai>=1.0.0
python-docx>=0.8.11
PyMuPDF>=1.22.5
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
graphviz>=0.20.1
python-dotenv>=1.0.0
pandas>=2.0.0
XlsxWriter>=3.0.3